Sample 1.3.0

+++ GamebaseSample/Assets/NhnCloud/GamebaseTools/SettingTool/Editor/Scripts/Util/ZipManager.cs
@@ -3,6 +3,8 @@ using System.Collections;
 using ICSharpCode.SharpZipLib.Zip;
 using System;
 using UnityEditor;
+using System.Text;
+using UnityEngine;
 
 namespace NhnCloud.GamebaseTools.SettingTool.Util
 {
@@ -17,20 +19,36 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
             UNKNOWN_ERROR,
         }
 
-        public static IEnumerator Extract(string zipFilePath, string unZipTargetFolderPath, Action<StateCode, string> callback, Action<FileStream> callbackFileStream = null, Action<float> progressCallback = null, string password = null, bool isDeleteZipFile = false)
+        public static IEnumerator Extract(string zipFilePath, string unZipTargetFolderPath, Action<StateCode, string> callback, Action<FileStream> callbackFileStream = null, Action<float> progressCallback = null, string password = null, bool isDeleteZipFile = false, bool isOverwrite = false)
         {
-#if UNITY_EDITOR_WIN
             string separatorPath = ReplaceDirectorySeparator(unZipTargetFolderPath);
-            string[] directorys = separatorPath.Split(Path.DirectorySeparatorChar);
-            string newPath = null;
-            string newPathRoot = null;
-            if (null != directorys && 0 < directorys.Length)
+            string[] directories = separatorPath.Split(Path.DirectorySeparatorChar);
+            string tempPath = null;
+            string tempPathRoot = null;
+            
+#if UNITY_EDITOR_WIN
+            if (null != directories && 0 < directories.Length)
             {
-                newPathRoot = directorys[0] + Path.DirectorySeparatorChar + "GamebaseSettingsToolTemp";
-                newPath = newPathRoot + Path.DirectorySeparatorChar + directorys[directorys.Length - 1];
+                tempPathRoot = directories[0] + Path.DirectorySeparatorChar + "GamebaseSettingsToolTemp";
+                tempPath = tempPathRoot + Path.DirectorySeparatorChar + directories[directories.Length - 1];
             }
 #elif UNITY_EDITOR_OSX
-		string newPath = unZipTargetFolderPath;
+            var sb = new StringBuilder();
+            if (null != directories && 0 < directories.Length)
+            {
+                for (int pathIndex = 0; pathIndex < directories.Length - 1; pathIndex++)
+                { 
+                    sb.Append(directories[pathIndex]);
+                    sb.Append(Path.DirectorySeparatorChar);
+                }
+
+                sb.Append("GamebaseSettingsToolTemp");
+                tempPathRoot = sb.ToString();
+
+                sb.Append(Path.DirectorySeparatorChar);
+                sb.Append(directories[directories.Length - 1]);
+                tempPath = sb.ToString();
+            }
 #endif
 
             if (true == string.IsNullOrEmpty(zipFilePath))
@@ -39,7 +57,7 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
                 yield break;
             }
 
-            if (true == string.IsNullOrEmpty(newPath))
+            if (true == string.IsNullOrEmpty(tempPath))
             {
                 callback(StateCode.FOLDER_PATH_NULL, unZipTargetFolderPath);
                 yield break;
@@ -55,14 +73,17 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
             FileStream fs;
             ZipInputStream zipInputStream;
             ZipEntry theEntry;
-            long nowCount = 0;
-            long totalCount = 0;
+            
+            long nowSize = 0;
+            long totalSize = 0;
             try
             {
                 zipInputStreamCount = new ZipInputStream(File.OpenRead(zipFilePath));
-                while (null != zipInputStreamCount.GetNextEntry())
+                
+                ZipEntry countEntry;
+                while ((countEntry = zipInputStreamCount.GetNextEntry()) != null)
                 {
-                    totalCount++;
+                    totalSize += countEntry.Size;
                 }
                 zipInputStreamCount.Close();
 
@@ -89,22 +110,20 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
                 string directoryName = Path.GetDirectoryName(theEntry.Name);
                 string fileName = Path.GetFileName(theEntry.Name);
 
-                Directory.CreateDirectory(newPath + "/" + directoryName);
+                Directory.CreateDirectory(tempPath + "/" + directoryName);
 
                 if (false == string.IsNullOrEmpty(fileName))
                 {
                     FileStream streamWriter = null;
                     try
                     {
-                        string filePath = Path.Combine(newPath, theEntry.Name);
+                        string filePath = Path.Combine(tempPath, theEntry.Name);
                         streamWriter = File.Create(filePath);
                     }
                     catch (Exception e)
                     {
                         zipInputStream.Close();
-#if UNITY_EDITOR_WIN
-                        FileUtil.DeleteFileOrDirectory(newPathRoot);
-#endif
+                        FileUtil.DeleteFileOrDirectory(tempPathRoot);
                         callback(StateCode.UNKNOWN_ERROR, e.Message);
                         yield break;
                     }
@@ -117,12 +136,13 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
                     int size = 2048;
                     byte[] data = new byte[2048];
 
+                    DateTime time = DateTime.Now;
                     while (true)
                     {
                         try
                         {
                             size = zipInputStream.Read(data, 0, data.Length);
-
+                            nowSize += size;
                             if (0 < size)
                             {
                                 streamWriter.Write(data, 0, size);
@@ -139,12 +159,21 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
                             callback(StateCode.UNKNOWN_ERROR, e.Message);
                             yield break;
                         }
+                        
+                        if ((DateTime.Now - time).TotalSeconds > 0.25f)
+                        {
+                            time = DateTime.Now;
+
+                            float progressStream = (float)nowSize / (float)totalSize;
+                            progressCallback(progressStream);
+                                
+                            yield return null;
+                        }
                     }
 
                     streamWriter.Close();
                 }
-                nowCount++;
-                float progress = (float)nowCount / (float)totalCount;
+                float progress = (float)nowSize / (float)totalSize;
                 progressCallback(progress);
                 yield return null;
             }
@@ -162,21 +191,27 @@ namespace NhnCloud.GamebaseTools.SettingTool.Util
                     callback(StateCode.UNKNOWN_ERROR, e.Message);
                 }
             }
-#if UNITY_EDITOR_WIN
             try
             {
-                if (true == Directory.Exists(unZipTargetFolderPath))
+                if(true == isOverwrite)
                 {
-                    FileUtil.DeleteFileOrDirectory(unZipTargetFolderPath);
+                    FileManager.CopyDirectory(tempPath, unZipTargetFolderPath, true);
                 }
-                FileUtil.MoveFileOrDirectory(newPath, unZipTargetFolderPath);
-                FileUtil.DeleteFileOrDirectory(newPathRoot);
+                else
+                {
+                    if (true == Directory.Exists(unZipTargetFolderPath))
+                    {
+                        FileUtil.DeleteFileOrDirectory(unZipTargetFolderPath);
+                    }
+                    FileUtil.MoveFileOrDirectory(tempPath, unZipTargetFolderPath);
+                }
+
+                FileUtil.DeleteFileOrDirectory(tempPathRoot);
             }
             catch (Exception e)
             {
                 callback(StateCode.UNKNOWN_ERROR, e.Message);
             }
-#endif
             callback(StateCode.SUCCESS, null);
 
         }
